requirements.txt
flask-restless==1.0.0b1
flask-sqlalchemy
requests
